ajustes aula 04

+++ 4. MLModel/src/core.py
@@ -38,7 +38,7 @@ class RawDataSchema(BaseModel):
     Data Model Input schema
     """
     tenure: int
-    MonthlyCharges : int
+    MonthlyCharges : float
     TotalCharges : float
     OnlineSecurity: str
     TechSupport : str
